extract handling unknown fields from NativeImagePointsToAnalysis into a dedicated class
